+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/default-o-table.datasource.ts
@@ -634,12 +634,12 @@ export class DefaultOTableDataSource extends DataSource<any> implements OTableDa
     return subGroups;
   }
 
-/**
- * Filters collapsed row group
- * @param data
- * @returns collapsed row group
- */
-filterCollapsedRowGroup(data: any): any[] {
+  /**
+   * Filters collapsed row group
+   * @param data
+   * @returns collapsed row group
+   */
+  filterCollapsedRowGroup(data: any): any[] {
     const self = this;
     return data.filter((row: any) => (row instanceof OTableGroupedRow) ? row.visible : this.belongsToExpandedGroupedRow(data, row));
   }