+++ CHANGELOG.md
@@ -8,7 +8,9 @@
   * Method `redirectLogin` has been removed in class `ServiceUtils`, now it is a method of Ontimize authentication implementation in `OntimizeAuthService`.
 
 ### Features
-* **o-table**: Option to group rows by column setted to true as default.
+* **o-table**: 
+  * Option to group rows by column setted to true as default.
+  * Collapse and expand same level when clicking on row group header.
 ## 8.2.5 (2021-05-26)
 ### Features
 * **o-image**: New attribute `max-file-size`. ([d8a84c0](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/d8a84c0)) Closes [#589](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/589)